backend/requirements.txt
python-dotenv
fastapi
pydantic
uvicorn
python-multipart
pandas
sentence-transformers
elasticsearch
qdrant-client
python-multipart
openai
ollama
anthropic
rerankers
